main/requirements.txt
colorama==0.4.6
pyfiglet==0.8.post1
pyyaml==6.0